Implemented Datepicker Components
The DatePickerOnStartPage component utilzies a Datepicker
component which was installed utilizing npm i vue3-datepicker

The Documentation for said package, can be found on:
https://icehaunter.github.io/vue3-datepicker/

It sports capacity to disable upcoming dates that are not eligable, 
disable dates on eligability criterias and is bound to be reactive
in terms of updating values.

Has a Minimum date of Today and Upper limit on Start date to be that of the end-date.