Integrated Dynamic Filters
In light of Filters needing to dynamically be able to update based on the state of the other factors - I had to manually map out every single state schematic in terms of different variables states - To cover all cases in terms of filters to process each situation.

The entire reason for this, was to make it so that every possible state was mapped - And full coverage across the state machine was reached - in regards to being able to cover all search cases.

One could re-factor the code and shorten it down - Or utilize more advanced javascript to boil down the repetition away - But this is a first draft of the setup, having all of the filters interact and depend on each other.